# api/openapi.yaml
openapi: 3.1.0
info:
  title: EVE-Cyno Data API
  summary: Deterministic EVE Online data, fit calculations and tools. No language model in the loop.
  description: |-
    Deterministic EVE Online data and calculations: community-fit search, fit breakdown and statistics (the Gofa dogma engine), SDE item search, and the tool layer behind the EVE-Cyno assistant, also reachable over MCP. Nothing here calls a language model; the same inputs give the same answers.

    ## Conventions

    - Errors are JSON, always: `{"error":{"code":"...","message":"..."}}`. The codes are stable (see `ErrorDetail`).
    - Every response carries `X-Request-ID`: your value when it is a token of 1-64 characters from `[A-Za-z0-9._-]`, otherwise a generated one.
    - Requests are rate limited per caller (a sliding window; the budget is stated on each operation): per API key when one is sent, otherwise per client IP. A refused request is `429` with `Retry-After`.
    - Request bodies are capped (256 KiB on the fit routes, 64 KiB on tools and MCP); an oversize body is `413`.
...

    ## Attribution and licences

    © 2014 Fenris Creations hf. All rights reserved. 'EVE', 'EVE Online', 'Fenris Creations', and all related logos and images are trademarks or registered trademarks of Fenris Creations hf. This material is used with limited permission of Fenris Creations. No official affiliation or endorsement by Fenris Creations is stated or implied. Created under the EVE Developer License Agreement.

...
servers:
...
paths:
...
  /tool/appraise_items:
    post:
      tags:
        - tools
      summary: Appraise a list of EVE Online items using Janice.
      description: |-
        Appraise a list of EVE Online items using Janice. Returns total buy/sell/split values in ISK and a link to the full appraisal.

        Sources: Janice.

        Access tier: **byo-key**. Runs only with your own Janice API key in the `X-Janice-Key` header (`400 janice_key_required` without it); the key is used for this request only and never logged or stored.

        Rate limit: 30 requests per minute per caller (per API key when one is sent, otherwise per client IP).
      operationId: appraise_items
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - name: X-Janice-Key
          in: header
          required: true
          description: Your own Janice API key (https://janice.e-351.com/). It is used for this request only, sent to Janice and nowhere else, and is never logged or stored by this service. This service never appraises with its own key.
          schema:
            type: string
            maxLength: 256
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: string
                  description: Items in multi-buy or EFT format, one per line with quantity, e.g. 'Tritanium x1000'
              required:
                - items
      responses:
        "200":
          description: The tool's answer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "413":
          $ref: '#/components/responses/PayloadTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
        "500":
          $ref: '#/components/responses/InternalError'
        "503":
          $ref: '#/components/responses/Unavailable'
        "504":
          $ref: '#/components/responses/Timeout'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
          description: Every non-2xx response has exactly this shape.
      required:
        - error
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          description: A stable, machine-readable code.
          enum:
            - not_found
            - method_not_allowed
            - invalid_request
            - payload_too_large
            - rate_limited
            - unavailable
            - upstream_error
            - unknown_tool
            - tool_disabled
            - api_key_required
            - invalid_api_key
            - key_not_permitted
            - janice_key_required
            - timeout
            - internal_error
        message:
          type: string
          description: A human-readable explanation; do not parse it.
      required:
        - code
        - message
...
    Source:
      type: object
      properties:
        name:
          type: string
          description: The upstream's name.
        url:
          type: string
          description: Where the upstream lives.
        license:
          type: string
          description: The terms the data is used under.
      required:
        - name
        - url
        - license
...
    ToolResponse:
      type: object
      properties:
        tool:
          type: string
          description: The tool that ran.
        version:
          type: string
          description: The build version of the service.
        result:
          type: object
          properties:
            text:
              type: string
              description: The answer as the chat assistant reads it.
            data:
              type: "null"
              description: The typed answer, or null when the tool has none or the call ended in a text-only answer (an unknown id, a parse failure); `text` then says why.
          required:
            - text
            - data
        attribution:
          type: array
          description: The upstreams behind the answer; attribution is a condition of several of their licences.
          items:
            $ref: '#/components/schemas/Source'
      required:
        - tool
        - version
        - result
        - attribution
...
  responses:
    BadRequest:
      description: 'The request is malformed: invalid JSON, or a missing or invalid parameter or header (`invalid_request`, `janice_key_required`).'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
...
    InternalError:
      description: The operation failed inside the service (`internal_error`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
...
    PayloadTooLarge:
      description: The request body is over the size cap (`payload_too_large`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
    RateLimited:
      description: The caller is over the rate limit; retry after the number of seconds in `Retry-After` (`rate_limited`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
    Timeout:
      description: The operation ran past its time budget (`timeout`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
    Unauthorized:
      description: The tool needs an API key and the request has none (`api_key_required`), or the key it sent is not valid (`invalid_api_key`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
    Unavailable:
      description: A dependency this operation needs is not loaded, for example the SDE or the fit corpus (`unavailable`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
...
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: An ID to correlate this request with the service's logs; echoed in the response.
      schema:
        type: string
        pattern: ^[A-Za-z0-9._-]{1,64}$
  headers:
    RequestID:
      description: 'The request''s ID: the caller''s `X-Request-ID` when it is a token of 1-64 characters from `[A-Za-z0-9._-]`, otherwise a generated one. Quote it when reporting a problem.'
      schema:
        type: string
    RetryAfter:
      description: Seconds until the caller's budget has room again.
      schema:
        type: integer
        minimum: 1
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      description: The same API key in the `X-API-Key` header, for clients that cannot set `Authorization`.
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      description: 'An API key as `Authorization: Bearer <key>`. Keyed tools need one; the other tools ignore it, but a key that is sent and not valid is refused (`401 invalid_api_key`).'
      scheme: bearer